🔧 chore(workflow_job): add conclusion label to the list of labels
What?

The goal of this change is to add `conclusion` to the labels list for workflow job metrics in github-exporter.

Why?

The `conclusion` label will contain the outcome of the workflow job, such as `success`, `failure`, `cancelled`, etc.

Context?

Today, in the labels list, we have the following default labels for workflow job metrics: owner, repo, name, title, branch, sha, identifier, run_id, run_attempt, labels, runner_id, runner_name, runner_group_id, runner_group_name, workflow_name.

The problem is that we don't know if the workflow job failed or not, we do have metrics, but the metrics are based on status:

- https://github.com/promhippie/github_exporter/blob/master/pkg/exporter/workflow_job.go#L145
- https://github.com/promhippie/github_exporter/blob/master/pkg/store/types.go#L68

The problem is that the jobs with conclusion: failure/success both have status completed (which is 4 https://github.com/promhippie/github_exporter/blob/master/pkg/exporter/workflow_job.go#L188), so for now we can't differentiate between a failed and a successful workflow job.

It is conclusion label that will help us to differentiate between a failed and a successful workflow job.