fix!: update billing github endpoints
Since GitHub has deprecated the previously used billing API we had to
change them as well. In combination with that we are more or less forced
to restructure the billing-related metrics to keep this exporter
codebase as simple as possible. All previous billing metrics are now
combined with the prefix `github_billing_usage_`.